+++ map/src/puck_group.rs
@@ -1,11 +1,8 @@
-use renderer::canvas_api::CanvasApi;
 use crate::mesh_loader::MeshLoader;
-use renderer::geometry_data::{GeometryData, ShapeData};
-use renderer::draw_commands::GeometryType;
-use renderer::mesh::mesh::StyledRangeInfo;
-use wgpu_canvas::MyCanvasApi;
+use wgpu_canvas::geometry_data::{GeometryData, GeometryType, ShapeData, StyledRangeInfo};
 use wgpu_canvas::render_group::RenderGroup;
 use wgpu_canvas::style_id::StyleId;
+use wgpu_canvas::MyCanvasApi;
 
 pub struct SimplePuck {}
