+++ map/src/route/route_group.rs
@@ -5,9 +5,7 @@ use glam::{DVec3, Vec2, Vec3};
 use lyon::geom::point;
 use lyon::lyon_tessellation::{LineCap, LineJoin};
 use lyon::path::Path;
-use renderer::draw_commands::{GeometryType, PolylineOptions};
-use renderer::geometry_data::{GeometryData, ShapeData, SvgData};
-use renderer::mesh::mesh::StyledRangeInfo;
+use wgpu_canvas::geometry_data::{GeometryData, GeometryType, PolylineOptions, ShapeData, StyledRangeInfo, SvgData};
 use wgpu_canvas::render_group::RenderGroup;
 use wgpu_canvas::style_id::StyleId;
 use wgpu_canvas::MyCanvasApi;
