+++ renderer/src/mesh/mesh.rs
@@ -5,11 +5,11 @@ use log::error;
 use lyon::lyon_tessellation::VertexBuffers;
 use std::ops::Range;
 use wgpu::{Buffer, RenderPass};
+use wgpu_canvas::geometry_data::StyledRangeInfo;
 use crate::buffer_pool::BufferPool;
 use crate::global_context::GlobalContext;
 
-#[derive(Clone)]
-pub struct StyledRangeInfo(pub u8, pub &'static str);
+
 #[derive(Clone)]
 pub struct StyledRange(pub Range<usize>, pub StyledRangeInfo);
