+++ renderer/src/mesh_layers/text_mesh_layer.rs
@@ -1,4 +1,4 @@
-use crate::geometry_data::{LineData, TextData};
+use wgpu_canvas::geometry_data::{LineData, TextData};
 use crate::global_context::GlobalContext;
 use crate::mesh_layers::BaseMeshLayer;
 use wgpu_canvas::render_modifier::SpatialData;
