+++ wgpu-canvas/Cargo.toml
@@ -8,3 +8,6 @@ wgpu = { workspace = true }
 strum = { workspace = true }
 glam = { workspace = true }
 geo-types = { workspace = true }
+lyon = { workspace = true}
+rustybuzz = "0.20.1"
+bytemuck = { workspace = true}
